refactor: rename @endo/chat-kit to @endo/spaces-util
Rename the shared message-rendering primitives package (markdown/value vnodes,
time formatters, chime, locator, token-autocomplete) from @endo/chat-kit to
@endo/spaces-util. The package is consumed by the chat host shell and every
space body (space-chat, space-channel), so the name reflects its role as the
shared substrate for chat spaces.

Moves packages/chat-kit → packages/spaces-util, updates the package name, all
22 import sites + dependency declarations across chat/space-chat/space-channel,
the tsconfig.json/typedoc.json excludes, and the migration docs. No behavior
change.

vite build passes; space-channel tsc 0; spaces-util-importing tests
(token-autocomplete, inbox-markdown) pass.
