+++ packages/chat/test/component/token-autocomplete-confined.test.js
@@ -6,7 +6,7 @@ import '@endo/init/debug.js';
 import test from 'ava';
 import { E } from '@endo/far';
 import { iterateReader } from '@endo/exo-stream/iterate-reader.js';
-import { tokenAutocompleteComponent } from '@endo/chat-kit/token-autocomplete.js';
+import { tokenAutocompleteComponent } from '@endo/spaces-util/token-autocomplete.js';
 import { makeMockPowers } from '../helpers/mock-powers.js';
 import { createDOM, createInputElements, tick } from '../helpers/dom-setup.js';
