+++ packages/chat/value-component.js
@@ -8,8 +8,8 @@ import harden from '@endo/harden';
 import { E } from '@endo/far';
 import { passStyleOf } from '@endo/pass-style';
 
-import { valueToVnodes } from '@endo/chat-kit/value-vnodes.js';
-import { MarkdownFragment } from '@endo/chat-kit/markdown-vnodes.js';
+import { valueToVnodes } from '@endo/spaces-util/value-vnodes.js';
+import { MarkdownFragment } from '@endo/spaces-util/markdown-vnodes.js';
 import {
   h,
   renderConfined,
