+++ packages/space-channel/src/channel-component.js
@@ -6,9 +6,9 @@ import harden from '@endo/harden';
 import { E } from '@endo/far';
 import { iterateReader } from '@endo/exo-stream/iterate-reader.js';
 
-import { timeFormatter, relativeTime } from '@endo/chat-kit/time-formatters.js';
-import { prepareTextWithPlaceholders } from '@endo/chat-kit/markdown-render.js';
-import { markdownToVnodes } from '@endo/chat-kit/markdown-vnodes.js';
+import { timeFormatter, relativeTime } from '@endo/spaces-util/time-formatters.js';
+import { prepareTextWithPlaceholders } from '@endo/spaces-util/markdown-render.js';
+import { markdownToVnodes } from '@endo/spaces-util/markdown-vnodes.js';
 import {
   Fragment,
   h,
@@ -422,7 +422,7 @@ export const channelComponent = async (
 
     const textWithPlaceholders = prepareTextWithPlaceholders(message.strings);
 
-    /** @type {import('@endo/chat-kit/markdown-vnodes.js').RenderToken} */
+    /** @type {import('@endo/spaces-util/markdown-vnodes.js').RenderToken} */
     const renderToken = index => {
       const edgeName = messageNames[index];
       if (edgeName === undefined) return null;
