+++ packages/space-chat/package.json
@@ -31,11 +31,11 @@
     "lint:eslint": "eslint '**/*.js'"
   },
   "dependencies": {
-    "@endo/chat-kit": "workspace:^",
     "@endo/daemon": "workspace:^",
     "@endo/exo-stream": "workspace:^",
     "@endo/far": "workspace:^",
     "@endo/harden": "workspace:^",
+    "@endo/spaces-util": "workspace:^",
     "preact": "^10.29.2"
   },
   "devDependencies": {
