+++ packages/spaces-util/src/locator.js
