+++ packages/spaces-util/src/token-autocomplete.js
