+++ packages/spaces-util/src/value-vnodes.js
